# laravel-bek/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patreon Clone API",
        "description": "API dokumentacija za platformu za podršku kreatorima",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/creators/{id}/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Javne objave kreatora",
                "description": "Display a listing of public posts for a specific creator.",
                "operationId": "e59e5272e70f9f76aa98393187da942a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kreatora",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Broj rezultata po stranici",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno učitane objave",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "objave": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostResource"
                                            }
                                        },
                                        "poruka": {
                                            "type": "string",
                                            "example": "Uspesno ucitane sve objave"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kreator nije pronađen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kreator ne postoji."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Kreiranje nove objave",
                "operationId": "9fdfb23b98422bf35855aed19967ed15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kreatora",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naslov",
                                    "sadrzaj",
                                    "pristup"
                                ],
                                "properties": {
                                    "naslov": {
                                        "type": "string",
                                        "example": "Nova objava"
                                    },
                                    "sadrzaj": {
                                        "type": "string",
                                        "example": "Sadržaj objave..."
                                    },
                                    "pristup": {
                                        "type": "string",
                                        "example": "javno",
                                        "enum": [
                                            "javno",
                                            "pretplatnici",
                                            "nivo"
                                        ]
                                    },
                                    "nivo_pristupa_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Objava kreirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Objava uspešno kreirana."
                                        },
                                        "post": {
                                            "$ref": "#/components/schemas/PostResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nema dozvolu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nemate dozvolu."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kreator ne postoji"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/my-posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista objava ulogovanog kreatora",
                "description": "Vraca sve objave ulogovanog korisnika koji ima status kreatora",
                "operationId": "6920c67c85d29924da9476d4014674e3",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista objava kreatora",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/PostResource"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Korisnik nije krator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Niste kreator."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "naslov": {
                        "type": "string",
                        "example": "Nova objava"
                    },
                    "sadrzaj": {
                        "type": "string",
                        "example": "Sadržaj objave..."
                    },
                    "datum_objave": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "pristup": {
                        "type": "string",
                        "example": "javno",
                        "enum": [
                            "javno",
                            "pretplatnici",
                            "nivo"
                        ]
                    },
                    "nivo_pristupa_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "creator": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "naziv_stranice": {
                                "type": "string",
                                "example": "Petrov kanal"
                            },
                            "user": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Petar Petrović"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}